requirements.txt
cython
numpy
scipy
sklearn
pandas
tensorflow-datasets>=4.2.0
# tensorflow-io
# tensorflow-text
tensorflow-addons>=0.11.1
setuptools>=47.1.1
librosa>=0.8.0
soundfile>=0.10.3
PyYAML>=5.3.1
matplotlib>=3.2.1
sox>=1.4.1
tqdm>=4.54.1
colorama>=0.4.4
nlpaug>=1.1.1
nltk>=3.5
sentencepiece>=0.1.94